Add files via upload
You are asked to ensure that the first and last names of people begin with a capital letter in their passports. 
For example, alison heck should be capitalised correctly as Alison Heck.


Given a full name, your task is to capitalize the name appropriately.

Input Format -
A single line of input containing the full name.

Constraints -
0<len(S)<100
The string consists of alphanumeric characters and spaces.